+++ src/reports/__init__.py
@@ -1,0 +1,2 @@
+import Configuration
+from ReportUtils import Reporter, runerror, get_configfile, get_template
