+++ src/reports/fife_reports/EfficiencyReporterPerVO.py
@@ -6,8 +6,7 @@ import datetime
 from elasticsearch_dsl import Search
 
 
-from . import Reporter, runerror, get_configfile, get_template
-from . import Configuration
+from . import Reporter, runerror, get_configfile, get_template, Configuration
 import reports.TextUtils as TextUtils
 import reports.NiceNum as NiceNum
