+++ src/reports/fife_reports/__init__.py
@@ -1,4 +1,1 @@
-from .. import Configuration
-
-# From the Reporter Module in the parent package, import various objects
-from ..Reporter import Reporter, runerror, get_configfile, get_template
+from .. import Configuration, Reporter, runerror, get_configfile, get_template
